+++ common/experiment/neutral/annot/three_scale.py
@@ -30,6 +30,8 @@ class ThreeScaleNeutralAnnotator(BaseNeutralAnnotator):
     # region private methods
 
     def __create_opinions_for_extraction(self, doc_id, data_type):
+        assert(isinstance(data_type, DataType))
+
         news = self._DocOps.read_news(doc_id=doc_id)
         opinions = self._OpinOps.read_etalon_opinion_collection(doc_id=doc_id)
         collection = self.__algo.make_neutrals(
@@ -52,13 +54,14 @@ class ThreeScaleNeutralAnnotator(BaseNeutralAnnotator):
 
         self.__algo = DefaultNeutralAnnotationAlgorithm(
             synonyms=self._DataIO.SynonymsCollection,
-            iter_parsed_news=lambda doc_id: self._DocOps.iter_parsed_news(doc_id, None),
-            iter_news_ids=self.iter_doc_ids_to_compare(),
+            iter_parsed_news=self._DocOps.iter_parsed_news(
+                doc_inds=self.iter_doc_ids_to_compare(),
+                frame_variant_collection=None),
             dist_in_terms_bound=data_io.DistanceInTermsBetweenOpinionEndsBound,
             ignored_entity_values=self.IGNORED_ENTITY_VALUES)
 
     def create_collection(self, data_type):
-        assert(isinstance(data_type, unicode))
+        assert(isinstance(data_type, DataType))
 
         filtered_iter = self.filter_non_created_doc_ids(data_type=data_type,
                                                         all_doc_ids=self.iter_doc_ids_to_compare())