+++ common/linked/text_opinions/collection.py
@@ -5,16 +5,13 @@ import cPickle as pickle
 
 from arekit.common.linked.text_opinions.wrapper import LinkedTextOpinionsWrapper
 from arekit.common.text_opinions.text_opinion import TextOpinion
-from arekit.common.parsed_news.collection import ParsedNewsCollection
 from arekit.common.text_opinions.collection import TextOpinionCollection
 
 
 logger = logging.getLogger(__name__)
 
 
-# TODO. Separate "Labeled": Create a different class that is
-# TODO. considered to receive LinkedTextOpinionCollection instance in __init__
-class LabeledLinkedTextOpinionCollection(TextOpinionCollection):
+class LinkedTextOpinionCollection(TextOpinionCollection):
     """
     Describes text opinions with a position precision and forward connection
 
@@ -25,22 +22,11 @@ class LabeledLinkedTextOpinionCollection(TextOpinionCollection):
 
     NO_NEXT_OPINION = None
 
-    # TODO. Parsed news collection (remove)
-    def __init__(self, parsed_news_collection):
-        # TODO. Parsed news collection (remove)
-        assert(isinstance(parsed_news_collection, ParsedNewsCollection))
-        # TODO. Parsed news collection (remove)
-        super(LabeledLinkedTextOpinionCollection, self).__init__(
-            parsed_news_collection=parsed_news_collection,
-            text_opinions=[])
+    def __init__(self):
+        super(LinkedTextOpinionCollection, self).__init__(text_opinions=[])
 
         # list describes that has i'th relation continuation in text.
         self.__next_opinion_id = []
-        # provides original label by text_opinion_id
-        # TODO. Relate to labeling
-        self.__text_opinion_labels = []
-        # labeling defined
-        self.__labels_defined = []
 
     def try_add_linked_text_opinions(self,
                                      linked_text_opinions,
@@ -75,79 +61,21 @@ class LabeledLinkedTextOpinionCollection(TextOpinionCollection):
                 registered_at_least_one = True
 
         if registered_at_least_one:
-            self.set_none_for_last_text_opinion()
+            self.__set_none_for_last_text_opinion()
 
         return discarded
 
-    # TODO. Private
-    def set_none_for_last_text_opinion(self):
+    def __set_none_for_last_text_opinion(self):
         self.__next_opinion_id[-1] = self.NO_NEXT_OPINION
 
     def register_text_opinion(self, text_opinion):
         assert(isinstance(text_opinion, TextOpinion))
-        super(LabeledLinkedTextOpinionCollection, self).register_text_opinion(text_opinion)
+        super(LinkedTextOpinionCollection, self).register_text_opinion(text_opinion)
         self.__next_opinion_id.append(text_opinion.TextOpinionID + 1)
-        # TODO. Relate to labeling
-        self.__text_opinion_labels.append(text_opinion.Sentiment)
-        self.__labels_defined.append(True)
 
     def remove_last_registered_text_opinion(self):
-        super(LabeledLinkedTextOpinionCollection, self).remove_last_registered_text_opinion()
+        super(LinkedTextOpinionCollection, self).remove_last_registered_text_opinion()
         del self.__next_opinion_id[-1]
-        # TODO. Relate to labeling
-        del self.__text_opinion_labels[-1]
-        del self.__labels_defined[-1]
-
-    # region public `check` methods
-
-    # TODO. Related to Labeling
-    def check_all_text_opinions_has_labels(self):
-        return not (False in self.__labels_defined)
-
-    # TODO. Related to Labeling
-    def check_all_text_opinions_without_labels(self):
-        return not (True in self.__labels_defined)
-
-    # endregion
-
-    # TODO. Related to Labeling
-    def apply_label(self, label, text_opinion_id):
-        assert(isinstance(text_opinion_id, int))
-
-        text_opinion = self[text_opinion_id]
-        assert(isinstance(text_opinion, TextOpinion))
-
-        if self.__labels_defined[text_opinion_id] is not False:
-            # assert(text_opinion.Sentiment == label)
-            if text_opinion.Sentiment != label:
-                logger.info("[Warning]: labels collision detected!")
-            return
-
-        text_opinion.set_label(label)
-        self.__labels_defined[text_opinion_id] = True
-
-    # region public `get` methods
-
-    # TODO. Remove
-    # TODO. Remove
-    # TODO. Remove
-    def get_labels_defined_count(self):
-        return self.__labels_defined.count(True)
-
-    # TODO. Remove
-    # TODO. Remove
-    # TODO. Remove
-    def get_original_label(self, text_opinion_id):
-        assert(isinstance(text_opinion_id, int))
-        return self.__text_opinion_labels[text_opinion_id]
-
-    # endregion
-
-    # TODO. Related to Labeling
-    def reset_labels(self):
-        for text_opinion in self:
-            text_opinion.set_label(self.__text_opinion_labels[text_opinion.TextOpinionID])
-        self.__labels_defined = [False] * len(self.__labels_defined)
 
     # region public serialization methods