+++ common/text_opinions/end_type.py
@@ -1,5 +1,0 @@
-class EntityEndType:
-
-    Source = u'source'
-
-    Target = u'target'