+++ common/text_opinions/helper.py
@@ -1,9 +1,10 @@
 from arekit.common.entities.base import Entity
+from arekit.common.labels.base import Label
+from arekit.common.opinions.base import Opinion
 from arekit.common.parsed_news.collection import ParsedNewsCollection
 from arekit.common.parsed_news.term_position import TermPosition, TermPositionTypes
-from arekit.common.text_opinions.collection import TextOpinionCollection
 from arekit.common.text_opinions.text_opinion import TextOpinion
-from arekit.common.text_opinions.end_type import EntityEndType
+from arekit.common.text_opinions.enums import EntityEndType, DistanceType
 
 
 # TODO. MOVE as written below
@@ -25,29 +26,38 @@ class TextOpinionHelper(object):
 
     # region public 'extract' methods
 
-    @staticmethod
-    def extract_entity_value(text_opinion, end_type):
-        parsed_news, id = TextOpinionHelper.__get(text_opinion, end_type)
-        return parsed_news.get_entity_value(id)
+    def extract_entity_value(self, text_opinion, end_type):
+        return self.__extract_entity_value(text_opinion=text_opinion,
+                                           end_type=end_type)
 
     def extract_entity_position(self, text_opinion, end_type, position_type=None):
-        return self.__get_entity_position_2(text_opinion=text_opinion,
-                                            end_type=end_type,
-                                            position_type=position_type)
+        return self.__get_entity_position(text_opinion=text_opinion,
+                                          end_type=end_type,
+                                          position_type=position_type)
 
-    @staticmethod
-    def extract_entity_sentence_index(text_opinion, end_type):
-        entity_position = TextOpinionHelper.__get_entity_position(text_opinion, end_type)
-        return entity_position.get_index(TermPositionTypes.SentenceIndex)
+    def to_opinion(self, text_opinion, label=None):
+        assert(isinstance(text_opinion, TextOpinion))
+        assert(isinstance(label, Label) or label is None)
+
+        source = self.__extract_entity_value(text_opinion=text_opinion,
+                                             end_type=EntityEndType.Source)
+
+        target = self.__extract_entity_value(text_opinion=text_opinion,
+                                             end_type=EntityEndType.Target)
+
+        return Opinion(source_value=source,
+                       target_value=target,
+                       sentiment=text_opinion.Sentiment if label is None else label)
 
     # endregion
 
     # region public 'calculate' methods
 
-    def calculate_distance_between_text_opinion_ends_in_terms(self, text_opinion):
+    def calc_dist_between_text_opinion_ends(self, text_opinion, distance_type):
         assert(isinstance(text_opinion, TextOpinion))
+        assert(isinstance(distance_type, DistanceType))
 
-        parsed_news = self.__parsed_news_collection.get_by_news_id(text_opinion)
+        parsed_news = self.__parsed_news_collection.get_by_news_id(news_id=text_opinion.NewsID)
 
         e1_id = self.__get_end_id(text_opinion=text_opinion,
                                   end_type=EntityEndType.Source)
@@ -57,28 +67,20 @@ class TextOpinionHelper(object):
 
         return TextOpinionHelper.__calc_distance(
             pos1=parsed_news.get_entity_position(id_in_document=e1_id),
-            pos2=parsed_news.get_entity_position(id_in_document=e2_id))
+            pos2=parsed_news.get_entity_position(id_in_document=e2_id),
+            position_type=DistanceType.to_position_type(distance_type))
 
-    def calculate_distance_between_entities_in_sentences(self, news_id, e1, e2):
+    def calc_dist_between_entities(self, news_id, e1, e2, distance_type):
         assert(isinstance(news_id, int))
         assert(isinstance(e1, Entity))
         assert(isinstance(e2, Entity))
+        assert(isinstance(distance_type, DistanceType))
 
         parsed_news = self.__parsed_news_collection.get_by_news_id(news_id)
         return TextOpinionHelper.__calc_distance(
             pos1=parsed_news.get_entity_position(e1.IdInDocument),
             pos2=parsed_news.get_entity_position(e2.IdInDocument),
-            position_type=TermPositionTypes.SentenceIndex)
-
-    def calculate_distance_between_entities_in_terms(self, news_id, e1, e2):
-        assert(isinstance(news_id, int))
-        assert(isinstance(e1, Entity))
-        assert(isinstance(e2, Entity))
-
-        parsed_news = self.__parsed_news_collection.get_by_news_id(news_id)
-        return TextOpinionHelper.__calc_distance(
-            pos1=parsed_news.get_entity_position(e1.IdInDocument),
-            pos2=parsed_news.get_entity_position(e2.IdInDocument))
+            position_type=DistanceType.to_position_type(distance_type))
 
     # endregion
 
@@ -87,13 +89,13 @@ class TextOpinionHelper(object):
     def check_ends_has_same_sentence_index(self, text_opinion):
         assert(isinstance(text_opinion, TextOpinion))
 
-        pos1 = self.__get_entity_position_2(text_opinion=text_opinion,
-                                            end_type=EntityEndType.Source,
-                                            position_type=TermPositionTypes.SentenceIndex)
+        pos1 = self.__get_entity_position(text_opinion=text_opinion,
+                                          end_type=EntityEndType.Source,
+                                          position_type=TermPositionTypes.SentenceIndex)
 
-        pos2 = self.__get_entity_position_2(text_opinion=text_opinion,
-                                            end_type=EntityEndType.Target,
-                                            position_type=TermPositionTypes.SentenceIndex)
+        pos2 = self.__get_entity_position(text_opinion=text_opinion,
+                                          end_type=EntityEndType.Target,
+                                          position_type=TermPositionTypes.SentenceIndex)
 
         return pos1 == pos2
 
@@ -114,25 +116,31 @@ class TextOpinionHelper(object):
         for ind_in_sent, term in it:
             if return_ind_in_sent:
                 yield ind_in_sent, term
-            else
+            else:
                 yield term
 
     # endregion
 
     # region private methods
 
-    def __get_entity_position_2(self, text_opinion, end_type, position_type=None):
-        end_id = TextOpinionHelper.__get_end_id(text_opinion, end_type)
-        parsed_news = self.__parsed_news_collection.get_by_news_id(text_opinion.NewsID)
+    def __extract_entity_value(self, text_opinion, end_type):
+        parsed_news, end_id = self.__get_related_news_and_end_id(text_opinion=text_opinion,
+                                                                 end_type=end_type)
+        return parsed_news.get_entity_value(end_id)
+
+    def __get_entity_position(self, text_opinion, end_type, position_type=None):
+        parsed_news, end_id = self.__get_related_news_and_end_id(text_opinion=text_opinion,
+                                                                 end_type=end_type)
         return parsed_news.get_entity_position(end_id, position_type)
 
-    # TODO. Should be removed.
-    # TODO. Should be removed.
-    # TODO. Should be removed.
-    @staticmethod
-    def __get_entity_position(text_opinion, end_type):
-        parsed_news, _id = TextOpinionHelper.__get(text_opinion, end_type)
-        return parsed_news.get_entity_position(_id)
+    def __get_related_news_and_end_id(self, text_opinion, end_type):
+        parsed_news = self.__get_related_news(text_opinion)
+        end_id = TextOpinionHelper.__get_end_id(text_opinion, end_type)
+        return parsed_news, end_id
+
+    def __get_related_news(self, text_opinion):
+        assert(isinstance(text_opinion, TextOpinion))
+        return self.__parsed_news_collection.get_by_news_id(text_opinion.NewsID)
 
     @staticmethod
     def __calc_distance(pos1, pos2, position_type=TermPositionTypes.IndexInDocument):
@@ -141,19 +149,6 @@ class TextOpinionHelper(object):
 
         return abs(pos1.get_index(position_type) -
                    pos2.get_index(position_type))
-
-    # TODO. Should be removed.
-    # TODO. Should be removed.
-    # TODO. Should be removed.
-    @staticmethod
-    def __get(text_opinion, end_type):
-        owner = text_opinion.Owner
-        assert(isinstance(owner, TextOpinionCollection))
-        id = TextOpinionHelper.__get_end_id(text_opinion, end_type)
-        pnc = owner.RelatedParsedNewsCollection
-        assert(isinstance(pnc, ParsedNewsCollection))
-        parsed_news = pnc.get_by_news_id(text_opinion.NewsID)
-        return parsed_news, id
 
     @staticmethod
     def __get_end_id(text_opinion, end_type):