+++ release_notes.md
@@ -1,12 +1,15 @@
 ## v0.20.5-rc
 
 Collection of parsed news become dispatched from text opinions collection.
+* Removed dependency from `RelatedParsedNewCollection` in TextOpinionCollection.
+* Labeling now separated from LinkedTextOpinion collection.
 
 Minor changes (light refactoring):
 
 * Bert now moved into separated folder from `contrib` directory.
 * frame_variants moved to `frames` directory.
 * Frame variants labeling in news now performed during `parse` operation.
+* `DataType` now enumeration. List of Supported data-types now a part of experiment
 
 ## v0.20.4-rc
 Updates: